made the hydrateOrders() action asynchronous in dashboard component
