+++ src/components/dashboard.jsx
@@ -13,23 +13,7 @@ const Dashboard = React.createClass({
     },
 
     componentDidMount () {
-        fetch('/api/orders')
-            .then(response => response.json())
-            .then(orders => {
-                const ordersToday = orders.map(orderWithStringDates => {
-                    return {
-                        ...orderWithStringDates,
-                        deadline: new Date(orderWithStringDates.deadline)
-                    };
-                }).filter(order => {
-                    const today = new Date();
-                    return today.getFullYear() === order.deadline.getFullYear() &&
-                        today.getMonth() === order.deadline.getMonth() &&
-                        today.getDate() === order.deadline.getDate();
-                });
-
-                this.props.dispatch(hydrateOrders(ordersToday));
-            });
+        this.props.dispatch(hydrateOrders());
     },
 
     _renderOrderTiles () {
